create package helper, clean code release 2 function
